Reuse joins related logic.
