+++ src/join.cpp
@@ -1409,3 +1409,271 @@ bool Join::merge_join_conditions(string& embedded_filter, string& query_filter) 
 
     return true;
 }
+
+void Join::aggregate_nested_references(single_filter_result_t *const reference_result,
+                                       reference_filter_result_t& ref_filter_result) {
+    // Add reference doc id in result.
+    auto temp_docs = new uint32_t[ref_filter_result.count + 1];
+    std::copy(ref_filter_result.docs, ref_filter_result.docs + ref_filter_result.count, temp_docs);
+    temp_docs[ref_filter_result.count] = reference_result->seq_id;
+
+    delete[] ref_filter_result.docs;
+    ref_filter_result.docs = temp_docs;
+    ref_filter_result.count++;
+    ref_filter_result.is_reference_array_field = false;
+
+    // Add references of the reference doc id in result.
+    auto& references = ref_filter_result.coll_to_references;
+    auto temp_references = new std::map<std::string, reference_filter_result_t>[ref_filter_result.count] {};
+    for (uint32_t i = 0; i < ref_filter_result.count - 1; i++) {
+        temp_references[i] = std::move(references[i]);
+    }
+    temp_references[ref_filter_result.count - 1] = std::move(reference_result->reference_filter_results);
+
+    delete[] references;
+    references = temp_references;
+}
+
+void Join::negate_left_join(id_list_t* const seq_ids, std::unique_ptr<uint32_t[]>& reference_docs, uint32_t& reference_docs_count,
+                            std::function<std::vector<uint32_t>(const uint32_t& ref_doc_id)> const& get_doc_id,
+                            const bool& is_match_all_ids_filter, std::vector<std::pair<uint32_t, uint32_t>>& id_pairs,
+                            std::set<uint32_t>& unique_doc_ids, negate_left_join_t& negate_left_join_info) {
+    uint32_t* negate_reference_docs = nullptr;
+    size_t negate_index = 0;
+    std::set<uint32_t> unique_negate_doc_ids;
+
+    // If the negate join is on all ids like !$CollName(id:*), we don't need to collect any references.
+    if (!is_match_all_ids_filter) {
+        auto it = seq_ids->new_iterator();
+        negate_reference_docs = new uint32_t[seq_ids->num_ids() - reference_docs_count];
+        for (size_t i = 0; i < reference_docs_count && it.valid(); i++) {
+            while (it.valid() && it.id() < reference_docs[i]) {
+                const auto &reference_doc_id = it.id();
+                it.next();
+
+                negate_reference_docs[negate_index++] = reference_doc_id;
+                std::vector<uint32_t> doc_ids = get_doc_id(reference_doc_id);
+                for (const auto &doc_id: doc_ids) {
+                    // If we have 3 products: product_a, product_b, product_c
+                    // and products_viewed like:
+                    // user_a:  [product_a]
+                    // user_b:  [product_a, product_b]
+                    // We should return product_b and product_c for "Products not seen by user_a".
+                    // So rejecting doc_id's already present in unique_doc_ids (product_a in the above example).
+                    if (doc_id == Index::reference_helper_sentinel_value || unique_doc_ids.count(doc_id) != 0) {
+                        continue;
+                    }
+
+                    id_pairs.emplace_back(doc_id, reference_doc_id);
+                    unique_negate_doc_ids.insert(doc_id);
+                }
+            }
+            if (!it.valid()) {
+                break;
+            }
+            while (i + 1 < reference_docs_count &&
+                   (reference_docs[i] + 1 == reference_docs[i + 1])) { // Skip consecutive ids.
+                i++;
+            }
+            it.skip_to(reference_docs[i] + 1);
+        }
+
+        if (reference_docs_count > 0 && it.valid()) {
+            it.skip_to(reference_docs[reference_docs_count - 1] + 1);
+        }
+        while (it.valid()) {
+            const auto &reference_doc_id = it.id();
+            it.next();
+
+            negate_reference_docs[negate_index++] = reference_doc_id;
+            std::vector<uint32_t> doc_ids = get_doc_id(reference_doc_id);
+            for (const auto &doc_id: doc_ids) {
+                // If we have 3 products: product_a, product_b, product_c
+                // and products_viewed like:
+                // user_a:  [product_a]
+                // user_b:  [product_a, product_b]
+                // We should return product_b and product_c for "Products not seen by user_a".
+                // So rejecting doc_id's already present in unique_doc_ids (product_a in the above example).
+                if (doc_id == Index::reference_helper_sentinel_value || unique_doc_ids.count(doc_id) != 0) {
+                    continue;
+                }
+
+                id_pairs.emplace_back(doc_id, reference_doc_id);
+                unique_negate_doc_ids.insert(doc_id);
+            }
+        }
+    }
+
+    reference_docs.reset(negate_reference_docs);
+    reference_docs_count = negate_index;
+
+    // Main purpose of `negate_left_join_info.excluded_ids` is help identify the doc_ids that don't have any references.
+    negate_left_join_info.excluded_ids_size = unique_doc_ids.size();
+    negate_left_join_info.excluded_ids.reset(new uint32_t[unique_doc_ids.size()]);
+    std::copy(unique_doc_ids.begin(), unique_doc_ids.end(), negate_left_join_info.excluded_ids.get());
+
+    unique_doc_ids = unique_negate_doc_ids;
+}
+
+void Join::get_ref_index_ids(num_tree_t& ref_index, const uint32_t& reference_doc_id,
+                             std::vector<std::pair<uint32_t, uint32_t>>& id_pairs,
+                             std::set<uint32_t>& unique_doc_ids, const bool& is_normal_join) {
+    size_t doc_ids_len = 0;
+    uint32_t* doc_ids = nullptr;
+
+    ref_index.search(NUM_COMPARATOR::EQUALS, reference_doc_id, &doc_ids, doc_ids_len);
+
+    for (size_t j = 0; j < doc_ids_len; j++) {
+        auto doc_id = doc_ids[j];
+        if (is_normal_join) {
+            id_pairs.emplace_back(doc_id, reference_doc_id);
+        }
+        unique_doc_ids.insert(doc_id);
+    }
+    delete[] doc_ids;
+}
+
+void Join::get_ref_index_ids(const spp::sparse_hash_map<uint32_t, int64_t, Hasher32>& ref_index,
+                             const uint32_t& reference_doc_id,
+                             std::vector<std::pair<uint32_t, uint32_t>>& id_pairs,
+                             std::set<uint32_t>& unique_doc_ids, const bool& is_normal_join) {
+    if (ref_index.count(reference_doc_id) == 0) { // Reference field might be optional.
+        return;
+    }
+    auto doc_id = ref_index.at(reference_doc_id);
+
+    if (doc_id == Index::reference_helper_sentinel_value) {
+        return;
+    }
+
+    if (is_normal_join) {
+        id_pairs.emplace_back(doc_id, reference_doc_id);
+    }
+    unique_doc_ids.insert(doc_id);
+}
+
+void Join::do_nested_join(std::function<std::pair<size_t, uint32_t*>(const uint32_t& ref_doc_id)> const& get_related_ids,
+                          const uint32_t& count, uint32_t const* const& reference_docs,
+                          filter_result_t* ref_filter_result, const std::string& ref_collection_name,
+                          filter_result_t& filter_result) {
+    // In case of nested join, we need to collect all the doc ids from the reference ids along with their references.
+    std::vector<std::pair<uint32_t, single_filter_result_t*>> id_pairs;
+    std::unordered_set<uint32_t> unique_doc_ids;
+
+    for (uint32_t i = 0; i < count; i++) {
+        auto& reference_doc_id = reference_docs[i];
+        auto reference_doc_references = std::move(ref_filter_result->coll_to_references[i]);
+
+        auto pair = get_related_ids(reference_doc_id);
+        size_t& doc_ids_len = pair.first;
+        uint32_t*& doc_ids = pair.second;
+
+        for (size_t j = 0; j < doc_ids_len; j++) {
+            auto doc_id = doc_ids[j];
+            auto reference_doc_references_copy = reference_doc_references;
+
+            id_pairs.emplace_back(std::make_pair(doc_id, new single_filter_result_t(reference_doc_id,
+                                                                                    std::move(reference_doc_references_copy),
+                                                                                    false)));
+            unique_doc_ids.insert(doc_id);
+        }
+
+        delete[] doc_ids;
+    }
+
+    if (id_pairs.empty()) {
+        return;
+    }
+
+    std::sort(id_pairs.begin(), id_pairs.end(), [](auto const& left, auto const& right) {
+        return left.first < right.first;
+    });
+
+    filter_result.count = unique_doc_ids.size();
+    filter_result.docs = new uint32_t[unique_doc_ids.size()];
+    filter_result.coll_to_references = new std::map<std::string, reference_filter_result_t>[unique_doc_ids.size()] {};
+
+    reference_filter_result_t previous_doc_references;
+    for (uint32_t i = 0, previous_doc = id_pairs[0].first + 1, result_index = 0; i < id_pairs.size(); i++) {
+        auto const& current_doc = id_pairs[i].first;
+        auto& reference_result = id_pairs[i].second;
+
+        if (current_doc != previous_doc) {
+            filter_result.docs[result_index] = current_doc;
+            if (result_index > 0) {
+                std::map<std::string, reference_filter_result_t> references;
+                references[ref_collection_name] = std::move(previous_doc_references);
+                filter_result.coll_to_references[result_index - 1] = std::move(references);
+            }
+
+            result_index++;
+            previous_doc = current_doc;
+            aggregate_nested_references(reference_result, previous_doc_references);
+        } else {
+            aggregate_nested_references(reference_result, previous_doc_references);
+        }
+    }
+
+    if (previous_doc_references.count != 0) {
+        std::map<std::string, reference_filter_result_t> references;
+        references[ref_collection_name] = std::move(previous_doc_references);
+        filter_result.coll_to_references[filter_result.count - 1] = std::move(references);
+    }
+
+    for (auto &item: id_pairs) {
+        delete item.second;
+    }
+
+    return;
+}
+
+void Join::process_related_ids(std::vector<std::pair<uint32_t, uint32_t>> id_pairs, const size_t& unique_doc_ids_size,
+                               const std::string& ref_collection_name, filter_result_t& filter_result) {
+    if (id_pairs.empty()) {
+        return;
+    }
+
+    std::sort(id_pairs.begin(), id_pairs.end(), [](auto const& left, auto const& right) {
+        return left.first < right.first;
+    });
+
+    filter_result = filter_result_t(unique_doc_ids_size, new uint32_t[unique_doc_ids_size],
+                                    new std::map<std::string, reference_filter_result_t>[unique_doc_ids_size] {});
+
+    std::vector<uint32_t> previous_doc_references;
+    for (uint32_t i = 0, previous_doc = id_pairs[0].first + 1, result_index = 0; i < id_pairs.size(); i++) {
+        auto const& current_doc = id_pairs[i].first;
+        auto const& reference_doc_id = id_pairs[i].second;
+
+        if (current_doc != previous_doc) {
+            filter_result.docs[result_index] = current_doc;
+            if (result_index > 0) {
+                auto& reference_result = filter_result.coll_to_references[result_index - 1];
+
+                auto r = reference_filter_result_t(previous_doc_references.size(),
+                                                   new uint32_t[previous_doc_references.size()],
+                                                   false);
+                std::copy(previous_doc_references.begin(), previous_doc_references.end(), r.docs);
+                reference_result[ref_collection_name] = std::move(r);
+
+                previous_doc_references.clear();
+            }
+
+            result_index++;
+            previous_doc = current_doc;
+            previous_doc_references.push_back(reference_doc_id);
+        } else {
+            previous_doc_references.push_back(reference_doc_id);
+        }
+    }
+
+    if (!previous_doc_references.empty()) {
+        auto& reference_result = filter_result.coll_to_references[filter_result.count - 1];
+
+        auto r = reference_filter_result_t(previous_doc_references.size(),
+                                           new uint32_t[previous_doc_references.size()],
+                                           false);
+        std::copy(previous_doc_references.begin(), previous_doc_references.end(), r.docs);
+        reference_result[ref_collection_name] = std::move(r);
+    }
+}
